+++ http4k-core/src/main/kotlin/org/http4k/websocket/internal.kt
@@ -1,5 +1,6 @@
 package org.http4k.websocket
 
+import org.http4k.core.Body
 import org.http4k.core.Request
 import org.http4k.websocket.WsStatus.Companion.NORMAL
 
@@ -24,4 +25,11 @@ abstract class PushPullAdaptingWebSocket(override val upgradeRequest: Request) :
     override fun onMessage(fn: (WsMessage) -> Unit) {
         messageHandlers.add(fn)
     }
+}
+
+class Http4kWebSocketAdapter(private val innerSocket: PushPullAdaptingWebSocket) {
+    fun onError(throwable: Throwable) = innerSocket.triggerError(throwable)
+    fun onClose(status: WsStatus) = innerSocket.triggerClose(status)
+
+    fun onMessage(body: Body) = innerSocket.triggerMessage(WsMessage(body))
 }