+++ http4k-serverless-lambda/integration-test/src/test/kotlin/org/http4k/aws/AwsLambdaApiClient.kt
@@ -106,3 +106,5 @@ data class FunctionDetails(val arn: String, val name: String)
 data class Role(val name: String)
 
 data class Region(val name: String)
+
+enum class LambdaIntegrationType { ApiGatewayV1, ApiGatewayV2, ApplicationLoadBalancer }