+++ http4k-serverless-lambda/integration-test/src/test/kotlin/org/http4k/serverless/lambda/client/ApiGatewayHttpClientTest.kt
@@ -28,9 +28,9 @@ abstract class ApiGatewayHttpClientTest(version: ApiIntegrationVersion) :
     HttpClientContract({ NoOpServerConfig }, client(version), client(version)) {
     override fun `connection refused are converted into 503`() = assumeTrue(false, "Unsupported client feature")
     override fun `handles response with custom status message`() = assumeTrue(false, "Unsupported client feature")
-    override fun `unknown host are converted into 503`()  = assumeTrue(false, "Unsupported client feature")
+    override fun `unknown host are converted into 503`() = assumeTrue(false, "Unsupported client feature")
 }
 
-class ApiGatewayV1ClientTest:ApiGatewayHttpClientTest(v1)
+class ApiGatewayV1ClientTest : ApiGatewayHttpClientTest(v1)
 
-class ApiGatewayV2ClientTest:ApiGatewayHttpClientTest(v2)
+class ApiGatewayV2ClientTest : ApiGatewayHttpClientTest(v2)