+++ http4k-serverless-lambda/integration-test/src/test/kotlin/org/http4k/serverless/lambda/client/LambdaHttpClient.kt
@@ -4,10 +4,13 @@ import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPEvent
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPResponse
-import org.http4k.aws.ApiIntegrationVersion
-import org.http4k.aws.ApiIntegrationVersion.v1
-import org.http4k.aws.ApiIntegrationVersion.v2
+import com.amazonaws.services.lambda.runtime.events.ApplicationLoadBalancerRequestEvent
+import com.amazonaws.services.lambda.runtime.events.ApplicationLoadBalancerResponseEvent
 import org.http4k.aws.FunctionName
+import org.http4k.aws.LambdaIntegrationType
+import org.http4k.aws.LambdaIntegrationType.ApiGatewayV1
+import org.http4k.aws.LambdaIntegrationType.ApiGatewayV2
+import org.http4k.aws.LambdaIntegrationType.ApplicationLoadBalancer
 import org.http4k.aws.Region
 import org.http4k.core.Body
 import org.http4k.core.Filter
@@ -23,10 +26,11 @@ import org.http4k.core.with
 import org.http4k.format.Jackson.auto
 import org.http4k.lens.BiDiBodyLens
 
-class LambdaHttpClient(functionName: FunctionName, region: Region, version: ApiIntegrationVersion) : Filter {
+class LambdaHttpClient(functionName: FunctionName, region: Region, version: LambdaIntegrationType) : Filter {
     private val adapter = when (version) {
-        v1 -> AwsClientV1HttpAdapter()
-        v2 -> AwsClientV2HttpAdapter()
+        ApiGatewayV1 -> AwsClientV1HttpAdapter()
+        ApiGatewayV2 -> AwsClientV2HttpAdapter()
+        ApplicationLoadBalancer -> AwsClientAlbHttpAdapter()
     }
 
     private fun callFunction(functionName: FunctionName) = Filter { next ->
@@ -96,7 +100,6 @@ internal class AwsClientV2HttpAdapter : AwsClientHttpAdapter<APIGatewayV2HTTPEve
 
     override fun invoke(request: Request) = APIGatewayV2HTTPEvent.builder()
         .withRawPath(request.uri.path)
-//        .withRawQueryString(request.uri.query)
         .withQueryStringParameters(request.uri.queries().toMap())
         .withBody(request.bodyString())
         .withHeaders(request.headers.toMap())
@@ -110,3 +113,27 @@ internal class AwsClientV2HttpAdapter : AwsClientHttpAdapter<APIGatewayV2HTTPEve
     override val requestLens = Body.auto<APIGatewayV2HTTPEvent>().toLens()
     override val responseLens = Body.auto<APIGatewayV2HTTPResponse>().toLens()
 }
+
+internal class AwsClientAlbHttpAdapter : AwsClientHttpAdapter<ApplicationLoadBalancerRequestEvent, ApplicationLoadBalancerResponseEvent> {
+    override fun invoke(response: ApplicationLoadBalancerResponseEvent) =
+        Response(Status(response.statusCode, ""))
+            .headers((response.headers
+                ?: emptyMap()).entries.fold(listOf(), { acc, next -> acc + (next.key to next.value) }))
+            .headers((response.multiValueHeaders ?: emptyMap()).entries.fold(listOf(), { acc, next ->
+                next.value.fold(acc, { acc2, next2 -> acc2 + (next.key to next2) })
+            }))
+            .body(response.body.orEmpty())
+
+
+    override fun invoke(request: Request) = ApplicationLoadBalancerRequestEvent().apply {
+        httpMethod = request.method.name
+        body = request.bodyString()
+        headers = request.headers.toMap()
+        path = request.uri.path
+        queryStringParameters = request.uri.query.toParameters().toMap()
+    }
+
+    override val requestLens = Body.auto<ApplicationLoadBalancerRequestEvent>().toLens()
+    override val responseLens = Body.auto<ApplicationLoadBalancerResponseEvent>().toLens()
+}
+