+++ http4k-serverless-lambda/integration-test/src/test/kotlin/org/http4k/serverless/lambda/client/client.kt
@@ -1,11 +1,11 @@
 package org.http4k.serverless.lambda.client
 
-import org.http4k.aws.ApiIntegrationVersion
-import org.http4k.aws.ApiIntegrationVersion.v1
 import org.http4k.aws.AwsApiGatewayApiClient
 import org.http4k.aws.AwsCredentialScope
 import org.http4k.aws.AwsCredentials
 import org.http4k.aws.AwsLambdaApiClient
+import org.http4k.aws.LambdaIntegrationType
+import org.http4k.aws.LambdaIntegrationType.ApiGatewayV1
 import org.http4k.aws.Region
 import org.http4k.aws.Role
 import org.http4k.client.JavaHttpClient
@@ -22,10 +22,10 @@ import org.http4k.serverless.lambda.inIntelliJOnly
 import org.junit.jupiter.api.Assumptions.assumeTrue
 
 val testFunctionClient by lazy {
-    testFunctionClient(v1)
+    testFunctionClient(ApiGatewayV1)
 }
 
-fun testFunctionClient(version: ApiIntegrationVersion) =
+fun testFunctionClient(version: LambdaIntegrationType) =
     LambdaHttpClient(DeployServerAsLambdaForClientContract.functionName(version), Config.region(awsConfig), version).then(awsClient("lambda"))
 
 val lambdaApiClient by lazy { AwsLambdaApiClient(awsClient("lambda"), Config.region(awsConfig)) }