+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate:
 
-### v3.266.0 (uncut)
+### v3.266.0
+- **http4k-*** : Upgrade some dependency versions.
 - **http4k-core** : Add CustomBasicAuth and ProxyBasicAuth filters. H/T @raymanoz for the PR
-- **http4k-core** : Implemented OriginPolicy for CORS. H/T @kratostaine  for the PR
+- **http4k-core** : Implemented OriginPolicy for CORS. H/T @kratostaine for the PR
+- **http4k-server-netty** : Websocket support added. H/T @carbotaniuman for the PR
 
 ### v3.265.0
 - **http4k-*** : Upgrade some dependency versions.